minor type changes

+++ src/prime_sieve.cpp
@@ -3,7 +3,7 @@
 // prime numbers less than some value n;
 // Author: Paul Stey
 // Date: 2017-03-14
-// CXXFLAGS: -Wall -O3 -std=c++11 
+// CXXFLAGS: -Wall -O3 -std=c++11
 
 #include <iostream>
 #include <vector>
@@ -18,13 +18,13 @@ std::vector<int> get_nums(int n) {
 
 std::vector<int> primes(int n) {
     std::vector<bool> prime_flag (n+1, true);
-    
+
     // Cross out non-primes by removing all multiples
     // of our existing primes. The incrementing of the
     // loop index `j` in the nested loop is the key.
-    for (int i = 2; i*i <= n; i++) {
+    for (size_t i = 2; i*i <= n; i++) {
     	if (prime_flag[i] == true) {
-            for (int j = i*2; j <= n; j += i) {
+            for (size_t j = i*2; j <= n; j += i) {
                 prime_flag[j] = false;
             }
         }
@@ -32,7 +32,7 @@ std::vector<int> primes(int n) {
 
     // collect the non-crossed out integers (i.e., primes)
     std::vector<int> res;
-    for (int i = 2; i <= n; i++) {
+    for (size_t i = 2; i <= n; i++) {
     	if (prime_flag[i] == true) {
     	    res.push_back(i);
     	}
